# backend/src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BeatRoute API",
    "version": "1.0.0",
    "description": "API for Spotify and Last.fm authentication, user data and playlists"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000"
    }
  ],
  "tags": [
    {
      "name": "Auth",
      "description": "Spotify authentication flow"
    },
    {
      "name": "Session",
      "description": "Application session endpoints"
    },
    {
      "name": "Me",
      "description": "Endpoints for the currently logged-in user"
    },
    {
      "name": "Last.fm",
      "description": "Last.fm account connection and user data"
    }
  ],
  "paths": {
    "/auth/spotify/login": {
      "get": {
        "tags": ["Auth"],
        "summary": "Redirect user to Spotify login",
        "responses": {
          "302": {
            "description": "Redirect to Spotify authorization page"
          }
        }
      }
    },
    "/auth/spotify/callback": {
      "get": {
        "tags": ["Auth"],
        "summary": "Handle Spotify OAuth callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect back to frontend"
          },
          "400": {
            "description": "OAuth error or state mismatch"
          }
        }
      }
    },
    "/auth/lastfm/login": {
      "get": {
        "tags": ["Last.fm"],
        "summary": "Redirect user to Last.fm authorization",
        "responses": {
          "302": {
            "description": "Redirect to the Last.fm authorization page"
          },
          "503": {
            "description": "Last.fm integration is not configured"
          }
        }
      }
    },
    "/auth/lastfm/callback": {
      "get": {
        "tags": ["Last.fm"],
        "summary": "Exchange a Last.fm authorization token for a session",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Last.fm session saved and user redirected to frontend"
          },
          "400": {
            "description": "Missing callback parameters or state mismatch"
          },
          "502": {
            "description": "Last.fm session could not be created"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Session"],
        "summary": "Check if user is logged in",
        "responses": {
          "200": {
            "description": "Authentication status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isLoggedIn": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Session"],
        "summary": "Log out current user",
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        }
      }
    },
    "/api/auth/lastfm": {
      "get": {
        "tags": ["Last.fm"],
        "summary": "Get Last.fm connection status",
        "responses": {
          "200": {
            "description": "Last.fm connection status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isConnected": {
                      "type": "boolean"
                    },
                    "username": {
                      "type": ["string", "null"]
                    }
                  },
                  "required": ["isConnected", "username"]
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/lastfm/logout": {
      "post": {
        "tags": ["Last.fm"],
        "summary": "Disconnect the Last.fm account",
        "responses": {
          "200": {
            "description": "Last.fm account disconnected"
          },
          "500": {
            "description": "Application session could not be updated"
          }
        }
      }
    },
    "/api/me/profile": {
      "get": {
        "tags": ["Me"],
        "summary": "Get current Spotify user profile",
        "responses": {
          "200": {
            "description": "Spotify user profile"
          },
          "401": {
            "description": "User not logged in"
          }
        }
      }
    },
    "/api/me/top-tracks": {
      "get": {
        "tags": ["Me"],
        "summary": "Get current user's top tracks",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "time_range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["short_term", "medium_term", "long_term"],
              "default": "medium_term"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Top tracks list"
          },
          "401": {
            "description": "User not logged in"
          }
        }
      }
    },
    "/api/me/top-artists": {
      "get": {
        "tags": ["Me"],
        "summary": "Get current user's top artists",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "time_range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["short_term", "medium_term", "long_term"],
              "default": "medium_term"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Top artists list"
          },
          "401": {
            "description": "User not logged in"
          }
        }
      }
    },
    "/api/lastfm/me": {
      "get": {
        "tags": ["Last.fm"],
        "summary": "Get the connected Last.fm user profile",
        "responses": {
          "200": {
            "description": "Last.fm user profile"
          },
          "401": {
            "description": "Last.fm account is not connected"
          },
          "502": {
            "description": "Last.fm profile could not be fetched"
          }
        }
      }
    },
    "/api/lastfm/track-info": {
      "get": {
        "tags": ["Last.fm"],
        "summary": "Get track metadata and genre candidates from Last.fm top tags",
        "parameters": [
          {
            "name": "artist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "track",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mbid",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Track metadata with Last.fm top tags",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LastfmTrackInfo"
                }
              }
            }
          },
          "400": {
            "description": "Neither MBID nor artist and track were provided"
          },
          "502": {
            "description": "Last.fm request failed"
          }
        }
      }
    },
    "/api/lastfm/artist-genres": {
      "post": {
        "tags": ["Last.fm"],
        "summary": "Build a genre distribution for Spotify top artists using Last.fm tags",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "artists": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 40,
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": ["artists"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Genre distribution among the supplied artists"
          },
          "400": {
            "description": "Invalid artist list"
          },
          "401": {
            "description": "Spotify session is missing or expired"
          },
          "502": {
            "description": "Last.fm request failed"
          }
        }
      }
    },
    "/api/lastfm/spotify-tracks/{spotifyTrackId}": {
      "get": {
        "tags": ["Last.fm"],
        "summary": "Resolve Spotify track metadata and Last.fm genre candidates",
        "parameters": [
          {
            "$ref": "#/components/parameters/SpotifyTrackId"
          }
        ],
        "responses": {
          "200": {
            "description": "Spotify metadata and Last.fm track information"
          },
          "401": {
            "description": "Spotify session is missing or expired"
          },
          "502": {
            "description": "Last.fm request failed"
          }
        }
      }
    },
    "/api/lastfm/spotify-tracks/{spotifyTrackId}/now-playing": {
      "post": {
        "tags": ["Last.fm"],
        "summary": "Set a Spotify track as now playing on Last.fm",
        "parameters": [
          {
            "$ref": "#/components/parameters/SpotifyTrackId"
          }
        ],
        "responses": {
          "200": {
            "description": "Now playing track accepted by Last.fm"
          },
          "401": {
            "description": "Spotify or Last.fm session is missing"
          },
          "502": {
            "description": "Spotify or Last.fm request failed"
          }
        }
      }
    },
    "/api/lastfm/spotify-tracks/{spotifyTrackId}/scrobble": {
      "post": {
        "tags": ["Last.fm"],
        "summary": "Scrobble a Spotify track on Last.fm",
        "parameters": [
          {
            "$ref": "#/components/parameters/SpotifyTrackId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "timestamp": {
                    "type": "integer",
                    "description": "UTC Unix timestamp when playback started"
                  },
                  "chosenByUser": {
                    "type": "boolean"
                  }
                },
                "required": ["timestamp"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Spotify track scrobble result"
          },
          "400": {
            "description": "Invalid timestamp or chosenByUser value"
          },
          "401": {
            "description": "Spotify or Last.fm session is missing"
          },
          "502": {
            "description": "Spotify or Last.fm request failed"
          }
        }
      }
    },
    "/api/lastfm/now-playing": {
      "post": {
        "tags": ["Last.fm"],
        "summary": "Set the currently playing track",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LastfmTrack"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Now playing track accepted by Last.fm"
          },
          "400": {
            "description": "Invalid track data"
          },
          "401": {
            "description": "Last.fm account is not connected"
          },
          "502": {
            "description": "Last.fm request failed"
          }
        }
      }
    },
    "/api/lastfm/scrobble": {
      "post": {
        "tags": ["Last.fm"],
        "summary": "Scrobble a track",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/LastfmTrack"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "timestamp": {
                        "type": "integer",
                        "description": "UTC Unix timestamp when playback started"
                      },
                      "chosenByUser": {
                        "type": "boolean"
                      }
                    },
                    "required": ["timestamp"]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scrobble result containing accepted and ignored counters"
          },
          "400": {
            "description": "Invalid track data"
          },
          "401": {
            "description": "Last.fm account is not connected or session expired"
          },
          "502": {
            "description": "Last.fm request failed"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "SpotifyTrackId": {
        "name": "spotifyTrackId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "LastfmTrack": {
        "type": "object",
        "properties": {
          "artist": {
            "type": "string"
          },
          "track": {
            "type": "string"
          },
          "album": {
            "type": "string"
          },
          "albumArtist": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "minimum": 1
          },
          "trackNumber": {
            "type": "integer",
            "minimum": 1
          },
          "mbid": {
            "type": "string"
          }
        },
        "required": ["artist", "track"]
      },
      "LastfmTrackInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": ["string", "null"]
          },
          "artist": {
            "type": ["string", "null"]
          },
          "mbid": {
            "type": ["string", "null"]
          },
          "url": {
            "type": ["string", "null"]
          },
          "genre": {
            "type": ["string", "null"],
            "description": "First Last.fm top tag; it is not an authoritative genre"
          },
          "genreCandidates": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "url": {
                  "type": ["string", "null"]
                }
              },
              "required": ["name", "url"]
            }
          },
          "genreSource": {
            "type": ["string", "null"],
            "enum": [
              "lastfm-top-tags",
              "lastfm-track-top-tags",
              "lastfm-artist-info-tags",
              null
            ]
          },
          "genreIsFallback": {
            "type": "boolean",
            "description": "True when tags come from the artist instead of the track"
          }
        },
        "required": [
          "name",
          "artist",
          "mbid",
          "url",
          "genre",
          "genreCandidates",
          "tags",
          "genreSource",
          "genreIsFallback"
        ]
      }
    }
  }
}